+++ experimental/packages/opentelemetry-instrumentation-http/src/http.ts
@@ -63,7 +63,7 @@ import { SEMATTRS_HTTP_ROUTE } from '@opentelemetry/semantic-conventions';
 /**
  * Http instrumentation instrumentation for Opentelemetry
  */
-export class HttpInstrumentation extends InstrumentationBase<Http> {
+export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentationConfig> {
   /** keep track on spans not ended */
   private readonly _spanNotEnded: WeakSet<Span> = new WeakSet<Span>();
   private _headerCapture;
@@ -94,29 +94,23 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
     );
   }
 
-  private _getConfig(): HttpInstrumentationConfig {
-    return this._config;
-  }
-
-  override setConfig(config?: HttpInstrumentationConfig): void {
+  override setConfig(config: HttpInstrumentationConfig = {}): void {
     super.setConfig(config);
     this._headerCapture = this._createHeaderCapture();
   }
 
   init(): [
-    InstrumentationNodeModuleDefinition<Https>,
-    InstrumentationNodeModuleDefinition<Http>,
+    InstrumentationNodeModuleDefinition,
+    InstrumentationNodeModuleDefinition,
   ] {
     return [this._getHttpsInstrumentation(), this._getHttpInstrumentation()];
   }
 
   private _getHttpInstrumentation() {
-    const version = process.versions.node;
-    return new InstrumentationNodeModuleDefinition<Http>(
+    return new InstrumentationNodeModuleDefinition(
       'http',
       ['*'],
-      moduleExports => {
-        this._diag.debug(`Applying patch for http@${version}`);
+      (moduleExports: Http): Http => {
         if (isWrapped(moduleExports.request)) {
           this._unwrap(moduleExports, 'request');
         }
@@ -143,9 +137,8 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
         );
         return moduleExports;
       },
-      moduleExports => {
+      (moduleExports: Http) => {
         if (moduleExports === undefined) return;
-        this._diag.debug(`Removing patch for http@${version}`);
 
         this._unwrap(moduleExports, 'request');
         this._unwrap(moduleExports, 'get');
@@ -155,12 +148,10 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
   }
 
   private _getHttpsInstrumentation() {
-    const version = process.versions.node;
-    return new InstrumentationNodeModuleDefinition<Https>(
+    return new InstrumentationNodeModuleDefinition(
       'https',
       ['*'],
-      moduleExports => {
-        this._diag.debug(`Applying patch for https@${version}`);
+      (moduleExports: Https): Https => {
         if (isWrapped(moduleExports.request)) {
           this._unwrap(moduleExports, 'request');
         }
@@ -187,9 +178,8 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
         );
         return moduleExports;
       },
-      moduleExports => {
+      (moduleExports: Https) => {
         if (moduleExports === undefined) return;
-        this._diag.debug(`Removing patch for https@${version}`);
 
         this._unwrap(moduleExports, 'request');
         this._unwrap(moduleExports, 'get');
@@ -312,7 +302,7 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
     startTime: HrTime,
     metricAttributes: MetricAttributes
   ): http.ClientRequest {
-    if (this._getConfig().requestHook) {
+    if (this.getConfig().requestHook) {
       this._callRequestHook(span, request);
     }
 
@@ -341,7 +331,7 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
           utils.getOutgoingRequestMetricAttributesOnResponse(responseAttributes)
         );
 
-        if (this._getConfig().responseHook) {
+        if (this.getConfig().responseHook) {
           this._callResponseHook(span, response);
         }
 
@@ -376,10 +366,10 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
 
           span.setStatus(status);
 
-          if (this._getConfig().applyCustomAttributesOnSpan) {
+          if (this.getConfig().applyCustomAttributesOnSpan) {
             safeExecuteInTheMiddle(
               () =>
-                this._getConfig().applyCustomAttributesOnSpan!(
+                this.getConfig().applyCustomAttributesOnSpan!(
                   span,
                   request,
                   response
@@ -473,13 +463,13 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
       if (
         utils.isIgnored(
           pathname,
-          instrumentation._getConfig().ignoreIncomingPaths,
+          instrumentation.getConfig().ignoreIncomingPaths,
           (e: unknown) =>
             instrumentation._diag.error('caught ignoreIncomingPaths error: ', e)
         ) ||
         safeExecuteInTheMiddle(
           () =>
-            instrumentation._getConfig().ignoreIncomingRequestHook?.(request),
+            instrumentation.getConfig().ignoreIncomingRequestHook?.(request),
           (e: unknown) => {
             if (e != null) {
               instrumentation._diag.error(
@@ -502,10 +492,10 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
 
       const spanAttributes = utils.getIncomingRequestAttributes(request, {
         component: component,
-        serverName: instrumentation._getConfig().serverName,
+        serverName: instrumentation.getConfig().serverName,
         hookAttributes: instrumentation._callStartSpanHook(
           request,
-          instrumentation._getConfig().startIncomingSpanHook
+          instrumentation.getConfig().startIncomingSpanHook
         ),
       });
 
@@ -531,10 +521,10 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
           context.bind(context.active(), request);
           context.bind(context.active(), response);
 
-          if (instrumentation._getConfig().requestHook) {
+          if (instrumentation.getConfig().requestHook) {
             instrumentation._callRequestHook(span, request);
           }
-          if (instrumentation._getConfig().responseHook) {
+          if (instrumentation.getConfig().responseHook) {
             instrumentation._callResponseHook(span, response);
           }
 
@@ -625,14 +615,14 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
       if (
         utils.isIgnored(
           origin + pathname,
-          instrumentation._getConfig().ignoreOutgoingUrls,
+          instrumentation.getConfig().ignoreOutgoingUrls,
           (e: unknown) =>
             instrumentation._diag.error('caught ignoreOutgoingUrls error: ', e)
         ) ||
         safeExecuteInTheMiddle(
           () =>
             instrumentation
-              ._getConfig()
+              .getConfig()
               .ignoreOutgoingRequestHook?.(optionsParsed),
           (e: unknown) => {
             if (e != null) {
@@ -656,7 +646,7 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
         hostname,
         hookAttributes: instrumentation._callStartSpanHook(
           optionsParsed,
-          instrumentation._getConfig().startOutgoingSpanHook
+          instrumentation.getConfig().startOutgoingSpanHook
         ),
       });
 
@@ -751,10 +741,10 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
       span.updateName(`${request.method || 'GET'} ${route}`);
     }
 
-    if (this._getConfig().applyCustomAttributesOnSpan) {
+    if (this.getConfig().applyCustomAttributesOnSpan) {
       safeExecuteInTheMiddle(
         () =>
-          this._getConfig().applyCustomAttributesOnSpan!(
+          this.getConfig().applyCustomAttributesOnSpan!(
             span,
             request,
             response
@@ -788,8 +778,8 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
      */
     const requireParent =
       options.kind === SpanKind.CLIENT
-        ? this._getConfig().requireParentforOutgoingSpans
-        : this._getConfig().requireParentforIncomingSpans;
+        ? this.getConfig().requireParentforOutgoingSpans
+        : this.getConfig().requireParentforIncomingSpans;
 
     let span: Span;
     const currentSpan = trace.getSpan(ctx);
@@ -832,7 +822,7 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
     response: http.IncomingMessage | http.ServerResponse
   ) {
     safeExecuteInTheMiddle(
-      () => this._getConfig().responseHook!(span, response),
+      () => this.getConfig().responseHook!(span, response),
       () => {},
       true
     );
@@ -843,7 +833,7 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
     request: http.ClientRequest | http.IncomingMessage
   ) {
     safeExecuteInTheMiddle(
-      () => this._getConfig().requestHook!(span, request),
+      () => this.getConfig().requestHook!(span, request),
       () => {},
       true
     );
@@ -863,7 +853,7 @@ export class HttpInstrumentation extends InstrumentationBase<Http> {
   }
 
   private _createHeaderCapture() {
-    const config = this._getConfig();
+    const config = this.getConfig();
 
     return {
       client: {