+++ experimental/packages/opentelemetry-instrumentation/src/instrumentation.ts
@@ -35,10 +35,11 @@ import {
 /**
  * Base abstract internal class for instrumenting node and web plugins
  */
-export abstract class InstrumentationAbstract<T = any>
-  implements Instrumentation
+export abstract class InstrumentationAbstract<
+  ConfigType extends InstrumentationConfig = InstrumentationConfig,
+> implements Instrumentation<ConfigType>
 {
-  protected _config: InstrumentationConfig;
+  protected _config: ConfigType;
 
   private _tracer: Tracer;
   private _meter: Meter;
@@ -48,7 +49,7 @@ export abstract class InstrumentationAbstract<T = any>
   constructor(
     public readonly instrumentationName: string,
     public readonly instrumentationVersion: string,
-    config: InstrumentationConfig = {}
+    config: ConfigType = {} as ConfigType // assuming ConfigType is an object with optional fields only
   ) {
     this._config = {
       enabled: true,
@@ -116,7 +117,7 @@ export abstract class InstrumentationAbstract<T = any>
    *
    * @returns an array of {@link InstrumentationModuleDefinition}
    */
-  public getModuleDefinitions(): InstrumentationModuleDefinition<T>[] {
+  public getModuleDefinitions(): InstrumentationModuleDefinition[] {
     const initResult = this.init() ?? [];
     if (!Array.isArray(initResult)) {
       return [initResult];
@@ -133,7 +134,7 @@ export abstract class InstrumentationAbstract<T = any>
   }
 
   /* Returns InstrumentationConfig */
-  public getConfig(): InstrumentationConfig {
+  public getConfig(): ConfigType {
     return this._config;
   }
 
@@ -141,7 +142,9 @@ export abstract class InstrumentationAbstract<T = any>
    * Sets InstrumentationConfig to this plugin
    * @param InstrumentationConfig
    */
-  public setConfig(config: InstrumentationConfig = {}): void {
+  public setConfig(config: ConfigType = {} as ConfigType): void {
+    // the assertion that {} is compatible with ConfigType may not be correct,
+    // ConfigType should contain only optional fields, but there is no enforcement in place for that
     this._config = Object.assign({}, config);
   }
 
@@ -172,7 +175,7 @@ export abstract class InstrumentationAbstract<T = any>
    * methods.
    */
   protected abstract init():
-    | InstrumentationModuleDefinition<T>
-    | InstrumentationModuleDefinition<T>[]
+    | InstrumentationModuleDefinition
+    | InstrumentationModuleDefinition[]
     | void;
 }